Semana 5, Enviar Correos, Llamadas telefonicas y por whatsapp, ver el correo y el mapa finalizados.

+++ app/src/main/java/com/example/lugares/model/Lugar.kt
@@ -24,7 +24,7 @@ data class Lugar(
     @ColumnInfo(name="latitud")
     val latitud: Double?,
     @ColumnInfo(name="altura")
-    val altura: Int?,
+    val altura: Double?,
     @ColumnInfo(name="rutaAudio")
     val rutaAudio: String?,
     @ColumnInfo(name="rutaImagen")
